+++ src/engine/kernel/norm/rms_norm.h
@@ -1,6 +1,6 @@
 #pragma once
 
-#include <cuda_fp16.h>
+#include <cuda_bf16.h>
 #include <cuda_runtime.h>
 
 #include "common/result.h"
@@ -8,8 +8,9 @@
 namespace ccinfer {
 namespace engine {
 
-Result<void> launch_rms_norm(const half* input, const half* weight, half* output, int rows, int dim,
-                              float eps, cudaStream_t stream);
+Result<void> launch_rms_norm(const __nv_bfloat16* input, const __nv_bfloat16* weight,
+                             __nv_bfloat16* output, int rows, int dim, float eps,
+                             cudaStream_t stream);
 
 }  // namespace engine
 }  // namespace ccinfer
